Updated test units
