+++ tests/test_data_has_retention_times.py
@@ -1,20 +1,23 @@
 import pytest
-import pandas as pd
 import os
 
 from .builders import SimpleDataBuilder
 from .builders import PandasDataBuilder
 from .builders import MatchMSDataBuilder
-from matchms.importing import load_from_msp
 
 here = os.path.abspath(os.path.dirname(__file__))
 testdata_dir_csv = os.path.join(here, 'data', 'csv')
 testdata_dir_msp = os.path.join(here, 'data', 'msp')
 
-@pytest.mark.parametrize("retention_times, expected",[((1.0, 2.0, 3.0), True), ((1.0, 2.0, -1), False)])
+
+@pytest.mark.parametrize("retention_times, expected", [
+    ((1.0, 2.0, 3.0), True),
+    ((1.0, 2.0, -1), False)
+])
 def test_simple_retention_times(retention_times, expected):
     data = SimpleDataBuilder().with_rt(retention_times).build()
     assert data.has_retention_times() == expected
+
 
 @pytest.mark.parametrize("filename_csv, expected", [
     (os.path.join(testdata_dir_csv, "has_retention_times_t.csv"), True),
@@ -24,9 +27,10 @@ def test_pandas_retention_times(filename_csv, expected):
     data = PandasDataBuilder().with_filename(filename_csv).build()
     assert data.has_retention_times() == expected
 
+
 @pytest.mark.parametrize("filename_msp, expected", [
      (os.path.join(testdata_dir_msp, "has_retention_times_t.msp"), True),
-     #(os.path.join(testdata_dir_msp, "has_retention_times_f.msp"), False)
+     (os.path.join(testdata_dir_msp, "has_retention_times_f.msp"), False)
 ])
 def test_matchms_retention_times(filename_msp, expected):
     data = MatchMSDataBuilder().with_filename(filename_msp).build()
